First pass at implementing write().
Doesn't work with the Go localfs, server -- we're getting an
unimplemented error.

I think this might be from expectSize(), in which case we should just
ignore it.

While working on this I realized that the way I've been checking the
flags for O_RDONLY is bogus, because O_RDONLY == 0 on Linux. The proper
mechanism is used in the new code, but I need to fix it elsewhere.